fix: class name play -> cache
