+++ docker-compose.yml
@@ -24,10 +24,7 @@ services:
     container_name: mysql
     image: mysql:8.0
     command: --default-authentication-plugin=mysql_native_password
-    #    volumes:
-    #      - ./mysql-data:/var/lib/mysql
-    # ports:
-    #   - "3308:3306"
+
     networks:
       - app-network
     restart: always
